# spec/api_contracts.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Virtual Try-On API",
    "description": "Comprehensive API specification for the Virtual Try-On application supporting avatar scanning, 3D product try-on, e-commerce, and user management",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@tryon.com",
      "url": "https://docs.tryon.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.tryon.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.tryon.com/v1",
      "description": "Staging server"
    },
    {
      "url": "https://dev-api.tryon.com/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/cart": {
      "get": {
        "tags": ["cart"],
        "summary": "Get shopping cart",
        "description": "Retrieve the user's current shopping cart with all items",
        "operationId": "getCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["cart"],
        "summary": "Clear cart",
        "description": "Remove all items from the shopping cart",
        "operationId": "clearCart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cart cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from login endpoint"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    },
    "schemas": {
...
      "CartResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Cart"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [true]
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Success message"
              }
            }
          }
        }
      },
...
      "Product": {
        "type": "object",
        "required": ["id", "name", "description", "price", "category"],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "originalPrice": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "default": "USD"
          },
          "brand": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "format": "uri"
                },
                "type": {
                  "type": "string",
                  "enum": ["front", "back", "side", "detail", "lifestyle"]
                },
                "resolution": {
                  "type": "string",
                  "enum": ["low", "medium", "high", "4k"]
                }
              }
            }
          },
          "model3D": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              },
              "format": {
                "type": "string",
                "enum": ["glb", "gltf", "fbx"]
              },
              "size": {
                "type": "string"
              },
              "polycount": {
                "type": "integer"
              }
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "size": {
                  "type": "string"
                },
                "color": {
                  "type": "string"
                },
                "colorHex": {
                  "type": "string",
                  "pattern": "^#[0-9A-Fa-f]{6}$"
                },
                "stock": {
                  "type": "integer",
                  "minimum": 0
                },
                "sku": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                }
              }
            }
          },
          "rating": {
            "type": "object",
            "properties": {
              "average": {
                "type": "number",
                "minimum": 0,
                "maximum": 5
              },
              "count": {
                "type": "integer",
                "minimum": 0
              }
            }
          },
          "tryonAvailable": {
            "type": "boolean",
            "default": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "material": {
            "type": "string"
          },
          "careInstructions": {
            "type": "string"
          },
          "returnPolicy": {
            "type": "string"
          },
          "shipping": {
            "type": "object",
            "properties": {
              "weight": {
                "type": "number"
              },
              "dimensions": {
                "type": "object",
                "properties": {
                  "length": {
                    "type": "number"
                  },
                  "width": {
                    "type": "number"
                  },
                  "height": {
                    "type": "number"
                  }
                }
              },
              "freeShippingThreshold": {
                "type": "number"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CartItem": {
        "type": "object",
        "required": ["id", "product", "quantity"],
        "properties": {
          "id": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "Cart": {
        "type": "object",
        "required": ["items", "total", "itemCount"],
        "properties": {
          "id": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "subtotal": {
            "type": "number",
            "minimum": 0
          },
          "tax": {
            "type": "number",
            "minimum": 0
          },
          "shipping": {
            "type": "number",
            "minimum": 0
          },
          "discount": {
            "type": "number",
            "minimum": 0
          },
          "total": {
            "type": "number",
            "minimum": 0
          },
          "itemCount": {
            "type": "integer",
            "minimum": 0
          },
          "couponCode": {
            "type": "string"
          },
          "estimatedDelivery": {
            "type": "string",
            "format": "date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}